requirements.txt
```
flask
psycopg2-binary
gunicorn
```
